requirements.txt
# Web框架
flask==2.3.3
fastapi==0.104.1
uvicorn==0.24.0

# 图数据库
neo4j==5.14.1

# 数据处理
pandas==2.1.3
numpy==1.24.3

# LLM相关
openai==1.3.7
transformers==4.35.2
torch>=2.0.1
accelerate>=0.24.1

# vLLM推理框架
vllm==0.2.1

# 工具类
requests==2.31.0
python-dotenv==1.0.0
pyyaml==6.0.1

# 日志
loguru==0.7.2dashscope==1.24.9
